requirements.txt
requests
python-telegram-bot
python-dotenv
pyTelegramBotAPI